requirements.txt
# Core runtime dependencies
torch>=2.1
numpy>=1.24
pyyaml>=6.0
gymnasium>=0.29
